Split peak_functions_c in two files.
peak_functions_c contained functions related with peak finding (as the
name suggests) AND functions related with sensor calibration (e.g,
removing pedestals, divinding by calibration constants). This commit
splits the original file, adding calib_sensor_functions_c, which
collects the sensor-calibration related functions. Tests files are
also split and the changes are propagated to base_cities and
peak_functions.
