+++ invisible_cities/reco/peak_functions_c.pxd
@@ -10,7 +10,7 @@ from scipy import signal
 """
 Computes the mean of a vector
 """
-cpdef double cmean(double [:] v)
+#cpdef double cmean(double [:] v)
 
 """
 Given a dictionary, pbounds, mapping potential peak number to potential peak, return a
@@ -47,11 +47,11 @@ pmt_active: a list of active PMTs
 n_MAU:  length of the MAU window
 thr_MAU: treshold above MAU to select sample
 """
-cpdef calibrated_pmt_sum(double [:, :] CWF,
-                         double [:] adc_to_pes,
-                         list       pmt_active = *,
-                         int        n_MAU      = *,
-                         double thr_MAU        = *)
+# cpdef calibrated_pmt_sum(double [:, :] CWF,
+#                          double [:] adc_to_pes,
+#                          list       pmt_active = *,
+#                          int        n_MAU      = *,
+#                          double thr_MAU        = *)
 
 
 """
@@ -64,11 +64,11 @@ pmt_active: a list of active PMTs
 n_MAU:  length of the MAU window
 thr_MAU: treshold above MAU to select sample
 """
-cpdef calibrated_pmt_mau(double [:, :]  CWF,
-                         double [:] adc_to_pes,
-                         list       pmt_active = *,
-                         int        n_MAU      = *,
-                         double     thr_MAU    = *)
+# cpdef calibrated_pmt_mau(double [:, :]  CWF,
+#                          double [:] adc_to_pes,
+#                          list       pmt_active = *,
+#                          int        n_MAU      = *,
+#                          double     thr_MAU    = *)
 
 
 """
@@ -128,77 +128,3 @@ contents expresses energy (e.g, in pes)
 The function returns a rebinned vector of T and E.
 """
 cpdef rebin_waveform(int ts, int t_finish, double[:] wf, int stride=*)
-
-
-"""Computes pedestal as average of the waveform"""
-cpdef sipm_subtract_baseline_and_normalize(np.ndarray[np.int16_t, ndim=2] sipm,
-                                           np.ndarray[np.float64_t, ndim=1] adc_to_pes)
-
-"""Computes pedetal using a MAU"""
-cpdef sipm_subtract_baseline_and_normalize_mau(np.ndarray[np.int16_t, ndim=2]sipm,
-                                               np.ndarray[np.float64_t, ndim=1] adc_to_pes,
-                                               int n_MAU=*)
-
-"""
-subtracts the baseline
-Uses a MAU to set the signal threshold (thr, in PES)
-"""
-cpdef sipm_signal_above_thr_mau(np.ndarray[np.int16_t, ndim=2] sipm,
-                                np.ndarray[np.float64_t, ndim=1] adc_to_pes,
-                                double thr,
-                                int n_MAU=*)
-
-"""
-subtracts the baseline
-Uses a MAU to set the signal threshold (thr, in PES)
-returns ZS waveforms for all SiPMs
-"""
-cpdef signal_sipm(np.ndarray[np.int16_t, ndim=2] SIPM,
-                  double [:] adc_to_pes, thr,
-                  int n_MAU=*, int Cal=*)
-
-"""
-Selects the SiPMs with signal
-and returns a dictionary:
-input: sipmzs[i,k], where:
-       i: runs over number of SiPms (with signal)
-       k: runs over waveform.
-       sipmzs[i,k] only includes samples above
-       threshold (e.g, dark current threshold)
-returns {j: [i, sipmzs[i]]}, where:
-       j: enumerates sipms with psum >0
-       i: sipm ID
-"""
-
-#\cpdef select_sipm(double [:, :] sipmzs)
-
-"""Given a dict with SIPMs (energies above threshold),
-return a dict of np arrays, where the key is the sipm
-with signal.
-
-input {j: [i, sipmzs[i]]}, where:
-       j: enumerates sipms with psum >0
-       i: sipm ID
-      S2d defining an S2 signal
-
-returns:
-      {i, sipm_i[i0:i1]} where:
-      i: sipm ID
-      sipm_i[i0:i1] waveform corresponding to SiPm i between:
-      i0: min index of S2d
-      i1: max index of S2d
-      only IF the total energy of SiPM is above thr
-
-"""
-#cpdef sipm_s2(dict dSIPM, dict S2, double thr)
-
-
-"""Given a vector with SIPMs (energies above threshold), and a
-dictionary of S2s, S2d, returns a dictionary of SiPMs-S2.  Each
-index of the dictionary correspond to one S2 and is a list of np
-arrays. Each element of the list is the S2 window in the SiPM (if
-not zero)
-
-"""
-#cpdef sipm_s2_dict(dict dSIPM, dict S2d, double thr)
-#cpdef sipm_s2_dict(double [:, :] sipmzs, dict S2d, double thr)
